+++ CHANGELOG.md
@@ -9,4 +9,5 @@ Version 2.0.0 (*In Progress*)
 - RxPreferences 2
 - Optional overload for RxPreferences instance
 - Rewrite in kotlin
-- 
+- Remove hard check on init context being an Application instance since we can retrieve the 
+application context defensively